backend/ai_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-multipart==0.0.9
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0
torch==2.1.2
